src/requirements.txt
Django==3.2.2
django-geojson==3.2.0
django-leaflet==0.28.0
psycopg2-binary==2.8.6
